Starting a splash page
This will allow you to select what "gender" you're interested in, then send you to the home page, which will be the swiping area. Leaving the API wrestling there for now.

Also began slices. The states "Dungeons" and "Dragons" will each hold 10 potential suitors for the user, selected at random. As the user swipes, the list will be repopulated with one entity per swipe, and the next swipe will create a random entity (via an API) and add it to the list.

Still need to make a Redux state for entities that are "swiped right" on, then work on basic styling and functionality as described.