Adding and removing matches functions as normal.
A comment from matches.js:

So removeMatch only checks names. That's pretty bad, but I can't find a way to compare the entire object that functions as expected (==,===,and Object.is all don't work). In a real application, this might compare usernames that must be unique or userIDs stored in a database.

I might implement something like a unique ID later if I have time.